ephy-profile-utils: migrate ephy_profile_utils_store/query_form_auth_data() to libsecret
We add a new SecretSchema that is specific to epiphany and intended
solely to store passwords for webforms. This is a better approach than
hacking the server url in order to store the names of the forms in it.

These methods are only used by EphyWebView to store the passwords and
to retrieve the password when there is a cache match and by one of the
early stages of password migration in the profile-migrator. If only this
patch is applied, it is likely that only newly saved patchs will work
properly, but others will remain intact.

https://bugzilla.gnome.org/show_bug.cgi?id=679918