ephy-profile-migrator: migrate form passwords to new schema
This completes the migration, by moving all passwords previously
stored as network passwords to the epiphany form passwords specific
schema. It must be noted that some of these passwords were not
properly stored as network passwords but as generic passwords, so a
throghout search was necessary in order to find all of them.

https://bugzilla.gnome.org/show_bug.cgi?id=679918