Introduced support for only 1 tab/window behaviour
For some situations like on the kiosk-mode, it's neccesary that
all the navigation is done on the same window and tab, which would
be opened fullscreen and without controls to switch between tabs,
so everything should be displated in the same view.

A new setting is introduced as part of the lockdown system to
enable this behaviour.

https://bugzilla.gnome.org/show_bug.cgi?id=703648